requirements.txt
asgiref==3.5.0
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
backports.zoneinfo==0.2.1
boto3==1.22.8
botocore==1.25.8
cffi==1.15.0
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
Django==4.0.2
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jmespath==1.0.0
Pillow==9.0.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2021.3
s3transfer==0.5.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.2.0
tzdata==2021.5
urllib3==1.26.9
zope.interface==5.4.0
